부대 단위 위치 배정 및 렌더 구현

+++ YaMang/YaMangServer/YaMangServer/ActionScheduler.h
@@ -6,7 +6,7 @@ struct ActionComparator
 {
 	bool operator()( const Action* lhs, const Action* rhs ) const
 	{
-		return (lhs->GetTime() > rhs->GetTime());
+		return ( lhs->GetTime() > rhs->GetTime() );
 	}
 };
 
@@ -18,13 +18,15 @@ public:
 	ActionScheduler();
 	~ActionScheduler();
 
-	void AddActionToScheduler( Action* addedAction, int64_t remainTime );
-	void DoScheduledAction();
+	void	AddActionToScheduler( Action* addedAction, int64_t remainTime );
+	void	DoScheduledAction();
 		
 private:
-	int64_t GetCurrentTick();
-	Clock::time_point m_BeginTime;
-	UINT m_CurrentTime=0;
-	std::priority_queue<Action*, std::vector<Action*>, ActionComparator> m_ActionQueue;
+	int64_t		GetCurrentTick();
+
+	Clock::time_point	m_BeginTime;
+	int64_t				m_CurrentTime = 0;
+
+	std::priority_queue<Action*, std::vector<Action*>, ActionComparator>	m_ActionQueue;
 };
